Add files via upload
With each purchase proceeds will b donated to kids in need all over the world in hopes of changing their situation and showing them that we care. Make a difference! Weather its Cancer societies, autism, special needs, homeless, or maybe the kids in your neighborhood we can make a difference for them one at a time. We care for EveryLife!